Made new test to pass

+++ src/com/trafficmon/AccountManager.java
@@ -1,12 +1,10 @@
 package com.trafficmon;
 
-import java.math.BigDecimal;
-
 // interface is for the ease of future change to library and adapter
 public interface AccountManager {
     void triggerInvestigationInto(Vehicle vehicle);
 
-    void deductCharge(Vehicle vehicle, BigDecimal charge) throws AccountNotRegisteredException, InsufficientCreditException;
+    void deductCharge(Vehicle vehicle, int charge) throws AccountNotRegisteredException, InsufficientCreditException;
 
-    void issuePenalty(Vehicle vehicle, BigDecimal charge);
+    void issuePenalty(Vehicle vehicle, int charge);
 }
